Extract tests for the helper module
